+++ mayan/apps/motd/tests/mixins.py
@@ -53,3 +53,41 @@ class MOTDTestMixin:
         self.test_message = Message.objects.create(
             label=TEST_LABEL, message=TEST_MESSAGE
         )
+
+
+class MessageViewTestMixin:
+    def _request_test_message_create_view(self):
+        pk_list = list(Message.objects.values_list('pk', flat=True))
+
+        response = self.post(
+            viewname='motd:message_create', data={
+                'label': TEST_LABEL,
+                'message': TEST_MESSAGE
+            }
+        )
+
+        self.test_message = Message.objects.exclude(
+            pk__in=pk_list
+        ).first()
+
+        return response
+
+    def _request_test_message_delete_view(self):
+        return self.post(
+            viewname='motd:message_single_delete', kwargs={
+                'message_id': self.test_message.pk
+            }
+        )
+
+    def _request_test_message_edit_view(self):
+        return self.post(
+            viewname='motd:message_edit', kwargs={
+                'message_id': self.test_message.pk
+            }, data={
+                'label': TEST_MESSAGE_EDITED,
+                'message': TEST_MESSAGE_EDITED
+            }
+        )
+
+    def _request_test_message_list_view(self):
+        return self.get(viewname='motd:message_list')