+++ HISTORY.rst
@@ -223,6 +223,7 @@
 - Add events to assets.
 - Re query search queryset after it has been sliced to workaround the ORM
   "Cannot filter a query once a slice has been taken".
+- Add events to the message of the day app.
 
 3.4.15 (2020-08-26)
 ===================